+++ instrumentation/opentelemetry-instrumentation-flask/src/opentelemetry/instrumentation/flask/__init__.py
@@ -251,7 +251,6 @@ from packaging import version as package_version
 import opentelemetry.instrumentation.wsgi as otel_wsgi
 from opentelemetry import context, trace
 from opentelemetry.instrumentation._semconv import (
-    _METRIC_ATTRIBUTES_SERVER_DURATION_NAME,
     _get_schema_url,
     _HTTPStabilityMode,
     _OpenTelemetrySemanticConventionStability,
@@ -268,6 +267,9 @@ from opentelemetry.instrumentation.propagators import (
 from opentelemetry.instrumentation.utils import _start_internal_or_server_span
 from opentelemetry.metrics import get_meter
 from opentelemetry.semconv.metrics import MetricInstruments
+from opentelemetry.semconv.metrics.http_metrics import (
+    HTTP_SERVER_REQUEST_DURATION,
+)
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.util.http import (
     get_excluded_urls,
@@ -282,6 +284,7 @@ _ENVIRON_SPAN_KEY = "opentelemetry-flask.span_key"
 _ENVIRON_ACTIVATION_KEY = "opentelemetry-flask.activation_key"
 _ENVIRON_REQCTX_REF_KEY = "opentelemetry-flask.reqctx_ref_key"
 _ENVIRON_TOKEN = "opentelemetry-flask.token"
+_ENVIRON_REQUEST_ROUTE_KEY = "request-route_key"
 
 _excluded_urls_from_env = get_excluded_urls("FLASK")
 
@@ -344,6 +347,11 @@ def _rewrapped_app(
                 excluded_urls is None
                 or not excluded_urls.url_disabled(flask.request.url)
             ):
+                if flask.request.url_rule:
+                    wrapped_app_environ[_ENVIRON_REQUEST_ROUTE_KEY] = str(
+                        flask.request.url_rule
+                    )
+
                 span = flask.request.environ.get(_ENVIRON_SPAN_KEY)
 
                 propagator = get_global_response_propagator()
@@ -386,6 +394,12 @@ def _rewrapped_app(
             duration_attrs_old = otel_wsgi._parse_duration_attrs(
                 attributes, _HTTPStabilityMode.DEFAULT
             )
+
+            if wrapped_app_environ.get(_ENVIRON_REQUEST_ROUTE_KEY, None):
+                duration_attrs_old[SpanAttributes.HTTP_ROUTE] = (
+                    wrapped_app_environ.get(_ENVIRON_REQUEST_ROUTE_KEY)
+                )
+
             duration_histogram_old.record(
                 max(round(duration_s * 1000), 0), duration_attrs_old
             )
@@ -393,6 +407,12 @@ def _rewrapped_app(
             duration_attrs_new = otel_wsgi._parse_duration_attrs(
                 attributes, _HTTPStabilityMode.HTTP
             )
+
+            if wrapped_app_environ.get(_ENVIRON_REQUEST_ROUTE_KEY, None):
+                duration_attrs_new[SpanAttributes.HTTP_ROUTE] = (
+                    wrapped_app_environ.get(_ENVIRON_REQUEST_ROUTE_KEY)
+                )
+
             duration_histogram_new.record(
                 max(duration_s, 0), duration_attrs_new
             )
@@ -553,7 +573,7 @@ class _InstrumentedFlask(flask.Flask):
         duration_histogram_new = None
         if _report_new(_InstrumentedFlask._sem_conv_opt_in_mode):
             duration_histogram_new = meter.create_histogram(
-                name=_METRIC_ATTRIBUTES_SERVER_DURATION_NAME,
+                name=HTTP_SERVER_REQUEST_DURATION,
                 unit="s",
                 description="measures the duration of the inbound HTTP request",
             )
@@ -684,7 +704,7 @@ class FlaskInstrumentor(BaseInstrumentor):
             duration_histogram_new = None
             if _report_new(sem_conv_opt_in_mode):
                 duration_histogram_new = meter.create_histogram(
-                    name=_METRIC_ATTRIBUTES_SERVER_DURATION_NAME,
+                    name=HTTP_SERVER_REQUEST_DURATION,
                     unit="s",
                     description="measures the duration of the inbound HTTP request",
                 )